+++ source/fsl_sdspi_disk/fsl_sdspi_disk.c
@@ -96,7 +96,7 @@ DRESULT sdspi_disk_ioctl(BYTE pdrv, BYTE cmd, void* buff)
         case GET_SECTOR_SIZE:
             if (buff)
             {
-                *(WORD *)buff = g_card.blockSize;
+                *(WORD *)buff = (WORD)g_card.blockSize;
             }
             else
             {
